improvement: move activities to activities package
- Move NotificationReceiver to receivers package
