+++ app/src/main/java/com/merive/press1mtimes/fragments/OptionsFragment.java
@@ -10,7 +10,7 @@ import android.widget.TextView;
 
 import androidx.fragment.app.Fragment;
 
-import com.merive.press1mtimes.MainActivity;
+import com.merive.press1mtimes.activities.MainActivity;
 import com.merive.press1mtimes.R;
 
 public class OptionsFragment extends Fragment {
