+++ app/src/main/java/com/merive/press1mtimes/fragments/SettingsFragment.java
@@ -10,7 +10,7 @@ import android.widget.TextView;
 import androidx.appcompat.widget.SwitchCompat;
 import androidx.fragment.app.Fragment;
 
-import com.merive.press1mtimes.MainActivity;
+import com.merive.press1mtimes.activities.MainActivity;
 import com.merive.press1mtimes.R;
 
 public class SettingsFragment extends Fragment {
