+++ app/src/main/java/com/merive/press1mtimes/fragments/SplashMessageFragment.java
@@ -12,7 +12,7 @@ import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
 
-import com.merive.press1mtimes.MainActivity;
+import com.merive.press1mtimes.activities.MainActivity;
 import com.merive.press1mtimes.R;
 
 public class SplashMessageFragment extends Fragment {
