+++ app/src/main/res/layout/activity_finish.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray"
-    tools:context=".FinishActivity">
+    tools:context=".activities.FinishActivity">
 
     <com.jetradarmobile.snowfall.SnowfallView
         android:id="@+id/coins"
